Rename slack.chatupdate to slack.chat
This way any helpers related to Slack's chat Web APIs can go here.
