+++ templatebot/slack/handlers/fileselect.py
@@ -4,7 +4,7 @@
 __all__ = ('handle_file_select_action',)
 
 from templatebot.slack.dialog import open_template_dialog
-from templatebot.slack.chatupdate import update_message
+from templatebot.slack.chat import update_message
 from .filedialogsubmission import render_template
