+++ templatebot/slack/handlers/projectselect.py
@@ -4,7 +4,7 @@
 __all__ = ('handle_project_select_action',)
 
 from templatebot.slack.dialog import open_template_dialog
-from templatebot.slack.chatupdate import update_message
+from templatebot.slack.chat import update_message
 
 
 async def handle_project_select_action(*, event_data, action_data, logger,
